backend/pyproject.toml
[build-system]
requires = ["setuptools>=70.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sql-trail"
version = "0.1.0"
description = "A web platform for SQL study."
readme = "./../README.md"
requires-python = ">=3.11.9"
license = {file = "LICENSE"}
authors = [
    {name = "Brenno Gomes Breda", email = "brenno.breda@icloud.com"}
]
dependencies = [
    "Flask >= 3.1.2",
    "flask-cors >= 6.0.1",
    "Flask-JWT-Extended >= 4.7.1",
    "python-dotenv >= 1.2.0",
    "SQLAlchemy >= 2.0.48",
    "PyMySQL >= 1.1.2",
    "supabase >= 2.22.2",
    "pandas >= 2.3.3",
    "pydantic >= 2.12.3",
    "requests >= 2.32.5",
]


[tool.delivery]
display_name = "SQL Trail"
company = "SQL Trail"


[project.optional-dependencies]
dev = [
    "ipython",
    "black",
    "flake8",
    "flask-debugtoolbar >= 0.15",
]


test = [
    "pytest >= 8.3",
    "pytest-cov >= 5.0",
    "pytest-flask",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]                  
namespaces = true                
exclude = ["tests*", "docs*"]    
